Split variation edit drawer into a dedicated VariationEditForm
Pulls the inline ProductEditForm out of ProductEdit into its own file
and adds a sibling VariationEditForm that the drawer picks when every
selected product is a variation. Shared field components, the
productFields registry, and DataForm pipeline are unchanged — both
forms still flow through the same merging/visibility/form-config
helpers, so updates to a shared field continue to benefit both views.

The split exists so variation-specific UX (parent inheritance
affordances, custom groupings, dedicated header bits) can evolve in
variation-edit-form.tsx without touching the regular product edit
form. No user-visible change in this commit.
